requirements.txt
requests
python-dotenv
influxdb